LibJS: Move CommonImplementations.h into Interpreter.cpp
Now that the Interpreter is the only user of these functions, we might
as well keep them in Interpreter.cpp which makes CLion less confused.

(cherry picked from commit ae0cfe4f2d260bfd804364c17f28494e3e8964c9)
